Refactor / use modern tooling, vite + react
