+++ .gitignore
@@ -1,6 +1,7 @@
 node_modules
 npm-debug.log
 dist
+dist-server
 public/bundle.js
 __pycache__/
 .DS_Store
