+++ Dockerfile
@@ -1,23 +1,10 @@
-# Use Node.js LTS version
-FROM node:18
+FROM node:22-slim
 
-# Create app directory
+ENV NODE_ENV=production
 WORKDIR /app
-
-# Copy package files
 COPY package*.json ./
-
-# Install dependencies
 RUN npm install
-
-# Copy source code
 COPY . .
-
-# Build the application
 RUN npm run build
-
-# Expose the port the app runs on
 EXPOSE 3000
-
-# Start the application
-CMD ["npm", "start"] 
+CMD ["npm", "start"]
