+++ README.md
@@ -1,1 +1,17 @@
 # pirate-bananagrams
+
+To start the dev server:
+
+```bash
+npm install
+npm run dev
+```
+
+To build and deploy:
+
+```bash
+npm run build
+npm start
+```
+
+Or use `Dockerfile` / `app.py`.
