+++ app.py
@@ -13,10 +13,11 @@ import modal
 app = modal.App("pirate-bananagrams")
 app.image = (
     modal.Image.from_registry("node:22-slim", add_python="3.13")
+    .env({"NODE_ENV": "production"})
     .add_local_file("package.json", "/app/package.json", copy=True)
     .add_local_file("package-lock.json", "/app/package-lock.json", copy=True)
     .run_commands("npm ci --prefix /app")
-    .add_local_dir("public", "/app/public")
+    .add_local_dir("dist-server", "/app/dist-server")
     .add_local_dir("dist", "/app/dist")
 )
 
@@ -29,4 +30,4 @@ app.image = (
 )
 @modal.web_server(port=3000)
 def web():
-    subprocess.Popen(["node", "dist/server/index.js"], cwd="/app")
+    subprocess.Popen(["node", "dist-server/index.js"], cwd="/app")
