+++ package.json
@@ -2,30 +2,35 @@
   "name": "pirate-bananagrams",
   "version": "1.0.0",
   "main": "index.js",
+  "type": "module",
   "scripts": {
-    "start": "node dist/server/index.js",
-    "build": "tsc && webpack",
-    "dev": "concurrently \"tsc -w\" \"webpack -w\" \"nodemon dist/server/index.js\""
+    "start": "node dist-server/index.js",
+    "build": "npm run build:client && npm run build:server",
+    "build:client": "vite build",
+    "build:server": "tsc -p tsconfig.server.json",
+    "dev": "concurrently \"vite\" \"tsx server/index.ts\""
   },
   "dependencies": {
-    "axios": "^1.8.4",
-    "express": "^4.21.2",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "socket.io": "^4.8.1",
-    "socket.io-client": "^4.4.1"
+    "express": "^5.1.0",
+    "socket.io": "^4.8.1"
   },
   "devDependencies": {
-    "@types/express": "^4.17.13",
-    "@types/node": "^16.11.7",
-    "@types/react": "^17.0.38",
-    "@types/react-dom": "^17.0.11",
+    "@types/express": "^5.0.1",
+    "@types/node": "^22.15.3",
+    "@types/react": "^19.1.2",
+    "@types/react-dom": "^19.1.2",
     "@types/socket.io": "^3.0.2",
-    "concurrently": "^7.0.0",
-    "nodemon": "^2.0.15",
-    "ts-loader": "^9.2.6",
-    "typescript": "^4.5.5",
-    "webpack": "^5.68.0",
-    "webpack-cli": "^4.9.2"
+    "@vitejs/plugin-react": "^4.4.1",
+    "concurrently": "^9.1.2",
+    "prettier": "^3.5.3",
+    "react": "^19.1.0",
+    "react-dom": "^19.1.0",
+    "socket.io-client": "^4.8.1",
+    "tsx": "^4.19.4",
+    "typescript": "~5.8.3",
+    "vite": "^6.3.3"
+  },
+  "engines": {
+    "node": ">=22"
   }
 }
