+++ src/dictionary.ts
