--- webpack.config.js
@@ -1,22 +1,0 @@
-const path = require('path');
-
-module.exports = {
-    entry: './src/client/index.tsx',
-    output: {
-        filename: 'bundle.js',
-        path: path.resolve(__dirname, 'public'),
-    },
-    module: {
-        rules: [
-            {
-                test: /\.tsx?$/,
-                use: 'ts-loader',
-                exclude: /node_modules/,
-            },
-        ],
-    },
-    resolve: {
-        extensions: ['.tsx', '.ts', '.js'],
-    },
-    mode: 'development',
-};
